Refactoring Options.
* Create an options package to hold file loading mode options.
* Move all options to a new file options.go for clarity
* Refactor table loading options
