+++ manifest_test.go
@@ -27,6 +27,7 @@ import (
 
 	"golang.org/x/net/trace"
 
+	"github.com/dgraph-io/badger/options"
 	"github.com/dgraph-io/badger/protos"
 	"github.com/dgraph-io/badger/table"
 	"github.com/dgraph-io/badger/y"
@@ -163,7 +164,7 @@ func TestOverlappingKeyRangeError(t *testing.T) {
 	lh0 := newLevelHandler(kv, 0)
 	lh1 := newLevelHandler(kv, 1)
 	f := buildTestTable(t, "k", 2)
-	t1, err := table.OpenTable(f, table.MemoryMap)
+	t1, err := table.OpenTable(f, options.MemoryMap)
 	require.NoError(t, err)
 	defer t1.DecrRef()
 
@@ -184,7 +185,7 @@ func TestOverlappingKeyRangeError(t *testing.T) {
 	lc.runCompactDef(0, cd)
 
 	f = buildTestTable(t, "l", 2)
-	t2, err := table.OpenTable(f, table.MemoryMap)
+	t2, err := table.OpenTable(f, options.MemoryMap)
 	require.NoError(t, err)
 	defer t2.DecrRef()
 	done = lh0.tryAddLevel0Table(t2)
